My_Project/backend/requirements.txt
Flask==2.2.5
requests==2.31.0
google-cloud-firestore==2.11.0
gunicorn==20.1.0
